org-program-main/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.6.0
sqlmodel>=0.0.16
sqlalchemy>=2.0.0
python-multipart>=0.0.9
lxml>=5.2.0
pytest>=7.4.0
httpx>=0.27.0
Pillow>=10.3.0
piexif>=1.1.3
